requirements.txt
streamlit
pymupdf
python-docx
python-dotenv
trafilatura
sentence-transformers
faiss-cpu
numpy
groq
pytest
pytest-mock

